requirements.txt
streamlit
ipykernel
networkx
openai==0.28
python-dotenv
matplotlib